fix: handle errors in user control form
When an error is triggered after submitting the form, the error
is not caught.

As such, no error toast is shown.

Let's fix this by adding a try..catch blog for the error when sending a
request.
If there is an error, then we show the error toast with the appropriate
message.